remove equip command, change pickup to handle equip.
